feature(enhancement): Enhancements and fixes on tests and controllers
Improved authorization checks for some endpoints
Other minor enhancements on models and services

+++ src/main/java/com/ninjaone/backendinterviewproject/controller/DeviceController.java
@@ -7,22 +7,21 @@ import com.ninjaone.backendinterviewproject.security.AuthenticatedUser;
 import com.ninjaone.backendinterviewproject.security.LoggedUser;
 import com.ninjaone.backendinterviewproject.service.DeviceBO;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/devices")
 public class DeviceController {
 
+	public static final String DEVICE_NOT_FOUND_OR_NOT_AUTHORIZED_ACCESS_MESSSAGE = "Device not found or you are not authorized to access it - deviceId=";
 	private final DeviceBO deviceBO;
 
 	public DeviceController(DeviceBO deviceBO) {
@@ -31,33 +30,52 @@ public class DeviceController {
 
 	@GetMapping(path = "/{id}")
 	public ResponseEntity<DeviceDTO> findById(@LoggedUser AuthenticatedUser authUser, @PathVariable("id") Long id) {
-		boolean noneMatch = authUser.getCustomer().getDevices().stream().noneMatch(d -> d.getId().equals(id));
-		if (noneMatch) {
-			throw new InvalidRequestException("Device not found or you are not authorized to access it - deviceId=" + id, null);
-		}
 		Device storedDevice = deviceBO.getById(id);
-		return ResponseEntity.ok(deviceBO.mapper().toDTO(storedDevice));
+		if (storedDevice.getCustomer().getId().equals(authUser.getId())) {
+			return ResponseEntity.ok(deviceBO.mapper().toDTO(storedDevice));
+		}
+		throw new InvalidRequestException(DEVICE_NOT_FOUND_OR_NOT_AUTHORIZED_ACCESS_MESSSAGE + id, null);
 	}
 
 	@GetMapping()
-	@PreAuthorize("hasAuthority('ADMIN')")
-	public ResponseEntity<List<DeviceDTO>> findAll() {
-		return ResponseEntity.ok(deviceBO.getAll().stream().map(deviceBO.mapper()::toDTO).collect(Collectors.toList()));
+	public ResponseEntity<List<DeviceDTO>> findAll(@LoggedUser AuthenticatedUser authUser) {
+		if (isAdmin(authUser)) {
+			List<DeviceDTO> allDevices = getAllDevices();
+			return ResponseEntity.ok(allDevices);
+		}
+		List<DeviceDTO> allDevicesFromAuthUser = getAllDevicesFrom(authUser);
+		return ResponseEntity.ok(allDevicesFromAuthUser);
 	}
 
 	@PutMapping()
-	public ResponseEntity<DeviceDTO> update(@LoggedUser AuthenticatedUser authUser, @RequestParam DeviceDTO deviceDTO) {
-		boolean noneMatch = authUser.getCustomer().getDevices().stream().noneMatch(d -> Objects.equals(d.getId(), deviceDTO.getId()));
-		if (noneMatch) {
-			throw new InvalidRequestException("Device not found or you are not authorized to access it - deviceId=" + deviceDTO.getId(), null);
+	public ResponseEntity<DeviceDTO> update(@LoggedUser AuthenticatedUser authUser, @RequestBody DeviceDTO deviceDTO) {
+		Device storedDevice = deviceBO.getById(deviceDTO.getId());
+		if (storedDevice.getCustomer().getId().equals(authUser.getId())) {
+			Device updatedDevice = deviceBO.update(deviceDTO);
+			return ResponseEntity.ok(deviceBO.mapper().toDTO(updatedDevice));
 		}
-		Device updatedDevice = deviceBO.update(deviceDTO);
-		return ResponseEntity.ok(deviceBO.mapper().toDTO(updatedDevice));
+		throw new InvalidRequestException(DEVICE_NOT_FOUND_OR_NOT_AUTHORIZED_ACCESS_MESSSAGE + deviceDTO.getId(), null);
 	}
 
 	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<Object> delete(@PathVariable("id") Long id) {
-		deviceBO.deleteById(id);
-		return ResponseEntity.noContent().build();
+	public ResponseEntity<Object> delete(@LoggedUser AuthenticatedUser authUser, @PathVariable("id") Long id) {
+		Device storedDevice = deviceBO.getById(id);
+		if (storedDevice.getCustomer().getId().equals(authUser.getId())) {
+			deviceBO.deleteById(id);
+			return ResponseEntity.noContent().build();
+		}
+		throw new InvalidRequestException(DEVICE_NOT_FOUND_OR_NOT_AUTHORIZED_ACCESS_MESSSAGE + id, null);
+	}
+
+	private boolean isAdmin(AuthenticatedUser authUser) {
+		return authUser.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN"));
+	}
+
+	private List<DeviceDTO> getAllDevices() {
+		return deviceBO.getAll().stream().map(deviceBO.mapper()::toDTO).collect(Collectors.toList());
+	}
+
+	private List<DeviceDTO> getAllDevicesFrom(AuthenticatedUser authUser) {
+		return deviceBO.findAllByCustomerId(authUser.getId()).stream().map(deviceBO.mapper()::toDTO).collect(Collectors.toList());
 	}
 }
